core: Add support for wider range of key codes
This patch changes the definition of KeyCode from an enum,
which was constraining the possible key codes, to a struct with
a u32 field, which covers all possible cases of a key code.

Key codes in Flash are not related to the physical buttons,
but rather to logical button values.
For instance, typing a non-ASCII key produces a key code related
to that character, and using a non-ANSI keyboard or a layout different
from US QWERTY also produces results matching that assumption.

+++ tests/input-format/src/format.rs
@@ -82,10 +82,10 @@ pub enum AutomatedEvent {
     },
 
     /// Press a key
-    KeyDown { key_code: u8 },
+    KeyDown { key_code: u32 },
 
     /// Release a key
-    KeyUp { key_code: u8 },
+    KeyUp { key_code: u32 },
 
     /// Input a character code
     TextInput { codepoint: char },
